# src/generated/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blawby API",
    "version": "1.0.0",
    "description": "Blawby Law Practice Management API with Stripe integration",
    "contact": {
      "name": "Blawby Support",
      "email": "support@blawby.com"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/users/{id}": {
      "get": {
        "summary": "Endpoint for /api/users/:id",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/users/:id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "summary": "Endpoint for /api/users/:id",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/users/:id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Endpoint for /api/users/:id",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/users/:id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/connected-accounts/{accountId}": {
      "get": {
        "summary": "Endpoint for /api/connected-accounts/:accountId",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/connected-accounts/:accountId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "accountId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/organizations/{organizationId}/onboarding-status": {
      "get": {
        "summary": "Endpoint for /api/organizations/:organizationId/onboarding-status",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/organizations/:organizationId/onboarding-status",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ]
}